+++ crates/entrypoint/src/cli/command/azure/entra/role/definition/mod.rs
@@ -1,9 +1,11 @@
 pub mod azure_entra_role_definition;
 pub mod azure_entra_role_definition_browse;
+pub mod azure_entra_role_definition_find;
 pub mod azure_entra_role_definition_list;
 
 pub use azure_entra_role_definition::AzureEntraRoleDefinitionCommand;
 pub use azure_entra_role_definition_browse::AzureEntraRoleDefinitionBrowseArgs;
+pub use azure_entra_role_definition_find::AzureEntraRoleDefinitionFindArgs;
 pub use azure_entra_role_definition_list::AzureEntraRoleDefinitionListArgs;
 use clap::Args;
 use eyre::Result;